+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- fix pub.dev warning
+- fix compatibility between `--subset` and `--sample`
+- support reverse complement the DNA sequence
+
 ## 0.0.1
 
 - support parsing multiline fasta format